Add RequestParamDefaultValue class
Added a new class named RequestParamDefaultValue. The main functionality of this new class is handling GET requests, as well as providing a default value if no parameters are given to its @RequestParam. It also contains an instance of the EconomicService class to fetch accounts by club ID, and optionally by name.